Bombs, blocks

+++ src/main/java/com/dayofpi/super_block_world/entity/custom/BoomBoomEntity.java
@@ -199,9 +199,9 @@ public class BoomBoomEntity extends Monster {
 
     private void dizzyEffect() {
         if (this.random.nextInt(6) == 0) {
-            double d0 = this.getX() - (double)this.getBbWidth() * Math.sin((double)(this.yBodyRot * ((float)Math.PI / 180F))) + (this.random.nextDouble() * 0.6D - 0.3D);
+            double d0 = this.getX() - (double)this.getBbWidth() * Math.sin(this.yBodyRot * ((float)Math.PI / 180F)) + (this.random.nextDouble() * 0.6D - 0.3D);
             double d1 = this.getY() + (double)this.getBbHeight() - 0.3D;
-            double d2 = this.getZ() + (double)this.getBbWidth() * Math.cos((double)(this.yBodyRot * ((float)Math.PI / 180F))) + (this.random.nextDouble() * 0.6D - 0.3D);
+            double d2 = this.getZ() + (double)this.getBbWidth() * Math.cos(this.yBodyRot * ((float)Math.PI / 180F)) + (this.random.nextDouble() * 0.6D - 0.3D);
             this.level().addParticle(ParticleTypes.ENTITY_EFFECT, d0, d1, d2, 0.4980392156862745D, 0.5137254901960784D, 0.5725490196078431D);
         }
     }
